enable BitmapSourceConverter and WriteableBitmapConverter for netcoreapp3.1
