+++ src/OpenCvSharp.Extensions/BitmapSourceConverter.cs
@@ -1,4 +1,4 @@
-﻿#if !DOTNETCORE
+﻿#if NET461 || NET48 || NETCOREAPP3_1
 using System;
 using System.Drawing;
 using System.Drawing.Imaging;
