+++ test/OpenCvSharp.Tests/extensions/BitmapSourceConverterTest.cs
@@ -1,5 +1,4 @@
-﻿#if DOTNET_FRAMEWORK 
-using System;
+﻿using System;
 using System.Globalization;
 using System.Runtime.InteropServices;
 using System.Windows;
@@ -11,12 +10,12 @@ namespace OpenCvSharp.Tests.Extensions
 {
     public class BitmapSourceConverterTest : TestBase
     {
-        [Fact]
+        [WindowsOnlyFact]
         public void BitmapSource8Bit()
         {
-            Scalar blueColor8 = new Scalar(200, 0, 0);
-            Scalar greenColor8 = new Scalar(0, 200, 0);
-            Scalar redColor8 = new Scalar(0, 0, 200);
+            var blueColor8 = new Scalar(200, 0, 0);
+            var greenColor8 = new Scalar(0, 200, 0);
+            var redColor8 = new Scalar(0, 0, 200);
 
             using (var mat = new Mat(1, 1, MatType.CV_8UC3, blueColor8))
             {
@@ -35,12 +34,12 @@ namespace OpenCvSharp.Tests.Extensions
             }
         }
 
-        [Fact]
+        [WindowsOnlyFact]
         public void BitmapSource16Bit()
         {
-            Scalar blueColor16 = new Scalar(32767, 0, 0);
-            Scalar greenColor16 = new Scalar(0, 32767, 0);
-            Scalar redColor16 = new Scalar(0, 0, 32767);
+            var blueColor16 = new Scalar(32767, 0, 0);
+            var greenColor16 = new Scalar(0, 32767, 0);
+            var redColor16 = new Scalar(0, 0, 32767);
 
             using (var mat = new Mat(1, 1, MatType.CV_16UC3, blueColor16))
             {
@@ -62,17 +61,17 @@ namespace OpenCvSharp.Tests.Extensions
         /// <summary>
         /// https://github.com/shimat/opencvsharp/issues/304
         /// </summary>
-        [StaFact(Skip = "sample")]
+        [WindowsOnlyStaFactAttribute(Skip = "sample")]
         public void BitmapSourceSample()
         {
             const int size = 250;
 
             BitmapSource bs8, bs16;
 
-            Scalar blueColor8 = new Scalar(128, 0, 0);
-            Scalar greenColor8 = new Scalar(0, 128, 0);
-            Scalar redColor8 = new Scalar(0, 0, 128);
-            Scalar whiteColor8 = new Scalar(255, 255, 255);
+            var blueColor8 = new Scalar(128, 0, 0);
+            var greenColor8 = new Scalar(0, 128, 0);
+            var redColor8 = new Scalar(0, 0, 128);
+            var whiteColor8 = new Scalar(255, 255, 255);
             using (var mat = new Mat(size, size, MatType.CV_8UC3, new Scalar(128, 128, 128)))
             {
                 mat.Rectangle(new Rect(15, 10, 100, 100), blueColor8, -1);
@@ -87,10 +86,10 @@ namespace OpenCvSharp.Tests.Extensions
                 bs8 = OpenCvSharp.Extensions.BitmapSourceConverter.ToBitmapSource(mat);
             }
 
-            Scalar blueColor16 = new Scalar(32767, 0, 0);
-            Scalar greenColor16 = new Scalar(0, 32767, 0);
-            Scalar redColor16 = new Scalar(0, 0, 32767);
-            Scalar whiteColor16 = new Scalar(65535, 65535, 65535);
+            var blueColor16 = new Scalar(32767, 0, 0);
+            var greenColor16 = new Scalar(0, 32767, 0);
+            var redColor16 = new Scalar(0, 0, 32767);
+            var whiteColor16 = new Scalar(65535, 65535, 65535);
             using (var mat = new Mat(size, size, MatType.CV_16UC3, new Scalar(32767, 32767, 32767)))
             {
                 mat.Rectangle(new Rect(15, 10, 100, 100), blueColor16, -1);
@@ -130,7 +129,7 @@ namespace OpenCvSharp.Tests.Extensions
         }
 
         private void AssertPixelValue<T>(Scalar expectedValue, BitmapSource bs)
-            where T : struct, IConvertible
+            where T : unmanaged
         {
             if (bs.PixelWidth != 1 || bs.PixelHeight != 1)
                 throw new Exception("1x1 image only");
@@ -147,4 +146,3 @@ namespace OpenCvSharp.Tests.Extensions
         }
     }
 }
-#endif
